feat: Handle images in CLI lock folder
Images are copied from permanent storage during edit so they are
previewable in the editor. Images are then copied back to permanent
storage when the edit is complete. The post text is updated between
these operations so it will have absolute URLs when stored but relative
URLs when edited.

Fixes Handle images in CLI lock folder #57